requirements.txt
fastapi
uvicorn
python-multipart
python-docx
pypdf
pytesseract
Pillow
